requirements.txt
asgiref==3.9.1
attrs==25.3.0
dj-database-url==3.0.1
Django==5.2.5
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg2-binary
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
uritemplate==4.2.0
